requirements.txt
opencv-python
mediapipe
numpy
pandas
scikit-learn
pyautogui
pillow

